u64

+++ source/m3_exec.h
@@ -1382,7 +1382,7 @@ d_m3Op (Const32) {
     if (!dest) {
         return m3Err_mallocFailed;  // o un altro codice di errore appropriato
     }
-
+    
     * (u32*) dest = value;
     nextOp();
 }
@@ -1392,7 +1392,7 @@ d_m3Op (Const64) {
     u64 value = MEMACCESS(u64, _mem, _pc);  
     _pc += 2;  // Su ESP32 sempre 2 perché M3_SIZEOF_PTR == 4
 
-    void* dest = (void*)m3SegmentedMemAccess(_mem, _sp + immediate(i32), sizeof(u32));
+    void* dest = (void*)m3SegmentedMemAccess(_mem, _sp + immediate(i32), sizeof(u64));
     
     if (!dest) {
         return m3Err_mallocFailed;  // o un altro codice di errore appropriato
